fix: usa global-error sin etiquetas html
